+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to this project will be documented in this file.
 - #349 AS3 - better handling of declarations
 - #735 AS3 - index out of bounds in deobfuscator 
 - AS3 deobfuscator on &&, || operators
+- Merged continues in try..catch
 
 ### Changed
 - AS3 test methods separated to classes